Create common Button component
