+++ components/Current/Delete.js
@@ -1,9 +1,8 @@
 import React from 'react';
-import { StyleSheet, TouchableOpacity } from 'react-native';
 import { useDispatch } from 'react-redux';
 
 import tasksModel from '../../data/store/tasks';
-import { Text } from '../common';
+import { Button } from '../common';
 
 export default function Delete(props) {
   if (!props.editing) return null;
@@ -12,27 +11,9 @@ export default function Delete(props) {
   const { deleteTask } = tasksModel.actions;
 
   return (
-    <TouchableOpacity
+    <Button
+      text="Delete"
       onPress={() => dispatch(deleteTask(props.id))}
-      style={styles.button}
-    >
-      <Text style={styles.text}>
-        Delete
-      </Text>
-    </TouchableOpacity>
+    />
   );
 }
-
-const styles = StyleSheet.create({
-  button: {
-    borderWidth: 1,
-    borderStyle: 'solid',
-    borderColor: 'gray',
-    borderRadius: 6,
-    padding: 8,
-  },
-
-  text: {
-    fontSize: 16,
-  },
-});
