+++ components/common/index.js
@@ -1,7 +1,9 @@
 import Text from './Text';
 import TextInput from './TextInput';
+import Button from './Button';
 
 export {
   Text,
   TextInput,
+  Button,
 };
